Update README.md
For some reason the create command didn't get the changes. However I found that using the command that i've mentioned, it stops the container, takes the changes, and also restarts the container. It only takes 5 seconds too. 

Maybe it's not optimal, so tell me if i am getting something wrong :)